requirements.txt
Flask
torch
torchvision
Pillow
fpdf
